requirements.txt
pyodbc
Flask==2.3.2
python-dotenv
google-cloud-storage==2.8.0
